my commit

+++ src/wpm.hpp
@@ -16,7 +16,7 @@ int calculate_wpm(float start_time){
         //average nikalna sakxas,sab wpm ko sum garera,store all the wpms in a vector
         //and push_back to push to the very vector and finally final wpm vanera dekhauna sakiyo,damnn
         const float end_time = GetTime();
-        float wpm = (float)3/(end_time-start_time) *(float) 60 ;    
+        float wpm = 3.0/(end_time-start_time) *60.0 ;    
         wpm_list_vector.push_back(wpm);
         
         int avg_wpm = accumulate(wpm_list_vector.begin(),wpm_list_vector.end(),0) / wpm_list_vector.size();
